+++ ElevateEngine/src/ElevateEngine/Renderer/OpenGL/OpenGLFrameBuffer.h
@@ -26,7 +26,7 @@ namespace Elevate
 		uint32_t m_TextureId;
 
 		// Size
-		float m_Width;
-		float m_Height;
+		uint32_t m_Width;
+		uint32_t m_Height;
 	};
 }
